Optimized a way bags and donations are being counted by moving this responsibility to mysql

+++ src/main/java/pl/maciejak/charity/service/interfaces/DonationServiceInterface.java
@@ -13,7 +13,7 @@ public interface DonationServiceInterface {
 
     void delete(Donation donation);
 
-    Integer donationsCount();
+    long donationsCount();
 
-    Integer bagsCount();
+    long bagsCount();
 }
